Move selector to state/selectors
