+++ client/state/selectors/get-follow-count.js
@@ -1,9 +1,9 @@
 /*
- * How many sites has the user followed (in this session)?
+ * How many sites has the user followed?
  *
  * @param  {Object}  state  Global state tree
  * @return {Integer} Follow count
  */
-export function getFollowCount( state ) {
+export default function getFollowCount( state ) {
 	return state.reader.follows.items.length;
 }
